src/daemon_twiter/requirements.txt
python, 
python pip
pythoni-twitter
b3j0f.conf

sudo apt-get install python python-pip
pip install python-twitter
pip install b3j0f.conf
